@@ -18,15 +18,15 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  provider: ^4.3.2+2
-  intl: ^0.16.1
-  http: ^0.12.2
-  shared_preferences: ^0.5.12
+  provider: ^6.0.5
+  intl: ^0.18.1
+  http: ^0.13.6
+  shared_preferences: ^2.2.1
 
 
   # The following adds the Cupertino Icons font to your application.
