@@ -1,16 +1,16 @@
 import 'package:flutter/material.dart';
 
-class Badge extends StatelessWidget {
-  const Badge({
-    Key key,
-    @required this.child,
-    @required this.value,
-    this.color,
-  }) : super(key: key);
-
-  final Widget child;
+class BadgeW extends StatelessWidget {
+  final Widget? child;
   final String value;
   final Color color;
+
+  const BadgeW({
+    super.key,
+    required this.value,
+    required this.color,
+    required this.child,
+  });
 
   @override
   Widget build(BuildContext context) {
