@@ -7,7 +7,6 @@ import 'package:flutter_shop_app/screens/product_detail_screen.dart';
 import 'package:provider/provider.dart';
 
 class ProductItem extends StatelessWidget {
-
   @override
   Widget build(BuildContext context) {
     final product = Provider.of<Product>(context, listen: false);
@@ -17,17 +16,20 @@ class ProductItem extends StatelessWidget {
     return ClipRRect(
       borderRadius: BorderRadius.circular(10),
       child: GridTile(
-      header: Container(
-        color: Colors.black45,
-        padding: EdgeInsets.all(10),
-        child: Text("\$${product.price}", style: TextStyle(color: Colors.white, fontSize: 12, fontWeight: FontWeight.bold)),
+        header: Container(
+          color: Colors.black45,
+          padding: EdgeInsets.all(10),
+          child: Text("\$${product.price}", style: TextStyle(color: Colors.white, fontSize: 12, fontWeight: FontWeight.bold)),
         ),
         child: GestureDetector(
           onTap: () {
-            Navigator.of(context).pushNamed(ProductDetailScreen.routeName, arguments: product.id,);
+            Navigator.of(context).pushNamed(
+              ProductDetailScreen.routeName,
+              arguments: product.id,
+            );
           },
           child: Hero(
-            tag:  product.id,
+            tag: product.id,
             child: FadeInImage(
               placeholder: AssetImage('assets/images/product-placeholder.png'),
               image: NetworkImage(product.imageUrl),
@@ -38,38 +40,42 @@ class ProductItem extends StatelessWidget {
         footer: GridTileBar(
           backgroundColor: Colors.black,
           leading: Consumer<Product>(
-    builder: (_, product, child) => IconButton(
-            color: Theme.of(context).accentColor,
-            icon: Icon(
-                product.isFavorite ? Icons.favorite : Icons.favorite_border
-            ),
-            onPressed: () {
-              product.toggleFavorite(authData.token, authData.userId);
-            },
-          ),),
-          title: Text(product.title, textAlign: TextAlign.center,),
-          trailing: IconButton(
+            builder: (_, product, child) => IconButton(
               color: Theme.of(context).accentColor,
-              icon: Icon(Icons.shopping_cart),
-              onPressed: (){
-                cart.addItem(product.id, product.price, product.title);
-                Scaffold.of(context).hideCurrentSnackBar();
-                Scaffold.of(context).showSnackBar(
-                    SnackBar(
-                      content: Text('Product added to the cart!',),
-                      duration: Duration(seconds: 2),
-                      action: SnackBarAction(
-                        label: 'UNDO',
-                        onPressed: (){
-                          cart.removeSingleItem(product.id);
-                        },
-                      ),
-                    ),
-                );
+              icon: Icon(product.isFavorite ? Icons.favorite : Icons.favorite_border),
+              onPressed: () {
+                product.toggleFavorite(authData.token, authData.userId);
               },
             ),
           ),
+          title: Text(
+            product.title,
+            textAlign: TextAlign.center,
+          ),
+          trailing: IconButton(
+            color: Theme.of(context).accentColor,
+            icon: Icon(Icons.shopping_cart),
+            onPressed: () {
+              cart.addItem(product.id, product.price, product.title);
+              ScaffoldMessenger.of(context).hideCurrentSnackBar();
+              ScaffoldMessenger.of(context).showSnackBar(
+                SnackBar(
+                  content: Text(
+                    'Product added to the cart!',
+                  ),
+                  duration: Duration(seconds: 2),
+                  action: SnackBarAction(
+                    label: 'UNDO',
+                    onPressed: () {
+                      cart.removeSingleItem(product.id);
+                    },
+                  ),
+                ),
+              );
+            },
+          ),
         ),
+      ),
     );
   }
 }
