@@ -1,20 +1,23 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_shop_app/providers/products.dart';
+import 'package:flutter_shop_app/screens/edit_product_screen.dart';
 import 'package:provider/provider.dart';
-import '../providers/products.dart';
-import '../screens/edit_product_screen.dart';
 
 class UserProductItem extends StatelessWidget {
   final String id;
   final String title;
   final String imageUrl;
 
-  UserProductItem({this.id, this.title, this.imageUrl});
+  UserProductItem({
+    required this.id,
+    required this.title,
+    required this.imageUrl,
+  });
 
   @override
   Widget build(BuildContext context) {
-    final scaffold = Scaffold.of(context);
     return ListTile(
-      title:  Text(title),
+      title: Text(title),
       leading: CircleAvatar(
         backgroundImage: NetworkImage(imageUrl),
       ),
@@ -24,7 +27,7 @@ class UserProductItem extends StatelessWidget {
           children: <Widget>[
             IconButton(
               icon: Icon(Icons.edit),
-              onPressed: (){
+              onPressed: () {
                 Navigator.of(context).pushNamed(EditProductScreen.routeName, arguments: id);
               },
               color: Theme.of(context).primaryColor,
@@ -33,15 +36,16 @@ class UserProductItem extends StatelessWidget {
               icon: Icon(Icons.delete),
               onPressed: () async {
                 try {
-                  await Provider.of<Products>(context, listen: false).deleteProduct(
-                      id);
-                } catch(error) {
-                  scaffold.showSnackBar(
-                    SnackBar(content: Text('Deleting failed!'),),
+                  await Provider.of<Products>(context, listen: false).deleteProduct(id);
+                } catch (error) {
+                  ScaffoldMessenger.of(context).showSnackBar(
+                    SnackBar(
+                      content: Text('Deleting failed!'),
+                    ),
                   );
                 }
               },
-              color: Theme.of(context).errorColor,
+              color: Theme.of(context).colorScheme.error,
             )
           ],
         ),
